fix: make reth-db compile without default features
